+++ example/output.tf
@@ -7,3 +7,8 @@ output "ironio_projects_example_names" {
   description = "The project names"
   value       = "${data.ironio_projects.example.names}"
 }
+
+output "ironio_queues_example_names" {
+  description = "The queue names"
+  value       = "${data.ironio_queues.example.names}"
+}