+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 FEATURES:
 
 * **New Data Source:** `ironio_projects`
+* **New Data Source:** `ironio_queues`
 * **New Resource:** `ironio_project`
 * **New Resource:** `ironio_pull_queue`
 * **New Resource:** `ironio_push_queue`